+ GetWordMetadata json input
