+++ My1kWordsEe/Services/Ai/OpenAiClient.cs
@@ -1,9 +1,6 @@
 ﻿using System.Text.Json;
-using System.Text.Json.Serialization;
 
 using CSharpFunctionalExtensions;
-
-using My1kWordsEe.Models;
 
 using OpenAI.Chat;
 
@@ -111,85 +108,5 @@ namespace My1kWordsEe.Services
                 MaxTokens = 400,
             });
         }
-
-        public static async Task<Result<SampleWord>> GetWordMetadata(
-            this OpenAiClient openAiClient,
-            string eeWord,
-            string? comment = null)
-        {
-            const string prompt =
-                "Teie sisend on eestikeelne sõna (ja selle sõna valikuline selgitus).\n" +
-                "Kui antud sõna ei ole eestikeelne, tagasta 404\n" +
-                "Teie väljund on sõna metaandmed JSON-is vastavalt antud lepingule:\n" +
-                "```\n{\n" +
-                "\"ee_word\": \"<antud sõna>\",\n" +
-                "\"en_word\": \"<english translation>\",\n" +
-                "\"en_words\": [<array of alternative english translations if applicable>],\n" +
-                "\"en_explanation\": \"<explanation of the word meaning in english>\",\n" +
-                "\"ee_explanation\": \"<sõna tähenduse seletus eesti keeles>\"\n" +
-                "}\n```\n";
-
-            var response = await openAiClient.CompleteAsync(
-                prompt,
-                string.IsNullOrEmpty(comment)
-                    ? eeWord
-                    : $"{eeWord} ({comment})",
-                new ChatCompletionOptions
-                {
-                    ResponseFormat = ChatResponseFormat.JsonObject,
-                    Temperature = 0.333f
-                });
-
-            if (response.IsFailure)
-            {
-                return Result.Failure<SampleWord>(response.Error);
-            }
-
-            // could be ommited if we integrate an EE dictionary within the app
-            if (response.Value.Contains("404"))
-            {
-                return Result.Failure<SampleWord>("Not an Estonian word");
-            }
-
-            openAiClient.ParseJsonResponse<WordMetadata>(response).Deconstruct(
-                out bool _,
-                out bool isParsingError,
-                out WordMetadata wordMetadata,
-                out string parsingError);
-
-            if (isParsingError)
-            {
-                return Result.Failure<SampleWord>(parsingError);
-            }
-
-            return Result.Success(new SampleWord
-            {
-                EeWord = wordMetadata.EeWord,
-                EnWord = wordMetadata.EnWord,
-                EnWords = wordMetadata.EnWords,
-                EnExplanation = wordMetadata.EnExplanation,
-                EeExplanation = wordMetadata.EeExplanation,
-            });
-        }
-
-        private class WordMetadata
-        {
-            [JsonPropertyName("ee_word")]
-            public required string EeWord { get; set; }
-
-            [JsonPropertyName("en_word")]
-            public required string EnWord { get; set; }
-
-            [JsonPropertyName("en_explanation")]
-            public required string EnExplanation { get; set; }
-
-            [JsonPropertyName("ee_explanation")]
-            public required string EeExplanation { get; set; }
-
-            [JsonPropertyName("en_words")]
-            public required string[] EnWords { get; set; } = Array.Empty<string>();
-        }
     }
-
-
 }
